Removed permit reserve from TcpStealerApi
